adding redis features
